Extract usage handlers to own folder
