+++ src/usage/commandDescriptions.js
@@ -1,0 +1,10 @@
+
+module.exports = [
+    ['[tournaments]', 'list open tournaments'],
+    ['whoami', 'show who you are on Challonge'],
+    ['connect [<challonge_username>]', 'connect your Slack and Challonge accounts'],
+    ['disconnect', 'disconnect your Slack and Challonge accounts'],
+    ['signup', 'sign up for a tournament'],
+    ['next', 'list open matches in tournaments you are part of'],
+    ['help', 'show this information'],
+];
